messenger_api/requirements.txt
Django
djangorestframework
psycopg2-binary
django-probes
pillow
django-redis
django-debug-toolbar
redis
psycopg2
django-environ
boto3
django-storages
django-lint